fastapi/fasiapi-sqlalchemy/src/dependency/Security.py
from fastapi.security import OAuth2PasswordBearer
from jose import JWTError, jwt
from fastapi import Security, HTTPException, status

oauth2_scheme = OAuth2PasswordBearer(tokenUrl="token")

SECRET_KEY = "your_secret_key"
ALGORITHM = "HS256"

credentials_exception = HTTPException(
    status_code=status.HTTP_401_UNAUTHORIZED,
    detail="Could not validate credentials",
    headers={"WWW-Authenticate": "Bearer"},
)

def get_current_user(token: str = Security(oauth2_scheme)):
    try:
        payload = jwt.decode(token, SECRET_KEY, algorithms=[ALGORITHM])
        username: str = payload.get("sub")
        if username is None:
            raise credentials_exception
    except JWTError:
        raise credentials_exception
    return username

# @app.get("/protected-route")
# def protected_route(current_user: str = Depends(get_current_user)):
#     return {"user": current_user}